Commit remaining safety cleanup changes

+++ radio/src/telemetry/telemetry.cpp
@@ -467,7 +467,7 @@ void deregisterTelemetryQueue(TelemetryQueue* queue)
 static void pushDataToQueue(TelemetryQueue* queue, uint8_t* data, int length)
 {
   if (queue && queue->hasSpace(length)) {
-    for (uint8_t i = 0; i < length; i += 1) {
+    for (int i = 0; i < length; i += 1) {
       queue->push(data[i]);
     }
   }
